makefile: set deb architecture dynamically
Looking at
https://unix.stackexchange.com/questions/303434/how-to-specify-deb-package-is-platform-specific-but-for-any-platform
the "correct" way would be to set the architecture to "any"
and then let dpkg-gencontrol set it to the right value.

This looks very complicated for a simple package.

Instead, just set it dynamically to the right
value using sed.

Tested on amd64 and arm64.

Fixes https://github.com/wcbonner/GoveeBTTempLogger/issues/16