Make `gvh-titlemap.txt` allow spaces as a delimiter
`gvh-titlemap.txt` was expected to use only tab characters as a
delimiter between the Bluetooth address and corresponding title.
This was not explained in the README.md file and was not obvious
when visually inspecting the sample mapping file.

Support using spaces as a delimiter to be less confusing.  Also
update the README.md to more clearly specify the file format.